// frontend/lib/api/generated/model/menuDiaDetailOut.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SIGA-UNMSM API
 * API del Sistema Integral de Gestión de Almacén y Abastecimiento Alimentario — UNMSM. Módulos activos: Autenticación, Usuarios, Almacenes, Catálogo Nutricional.
 * OpenAPI spec version: 0.1.0
 */
import type { PlatoOut } from './platoOut';

export interface MenuDiaDetailOut {
  menu_dia_id: number;
  menu_id: number;
  fecha: string;
  tipo_servicio: string;
  raciones_programadas: number;
  platos: PlatoOut[];
}